requirements.txt
langchain
langchain_community
chainlit
langchain_groq
langchain_pinecone
py-dotenv
pypdf

